+++ frontend/src/utils/format.js
@@ -1,0 +1,3 @@
+export function formatMoney(value, currency = 'USD') {
+  return new Intl.NumberFormat('en-US', { style: 'currency', currency }).format(Number(value || 0))
+}